Renamed all SingleVideoViewer file names
and updated index.js and any corresponding import paths.
Fixed error in preload method - somehow ‘fetchImage’ was deleted?! Fat-finger delete?
